+++ Makefile.common
@@ -1122,11 +1122,12 @@ ifeq ($(HAVE_NETWORKING), 1)
 
    # Netplay
 	DEFINES += -DHAVE_NETWORK_CMD
-	OBJ += network/netplay/netplay_frontend.o \
+	OBJ += network/netplay/netplay_delta.o \
+			network/netplay/netplay_frontend.o \
+			network/netplay/netplay_handshake.o \
 			network/netplay/netplay_init.o \
 			network/netplay/netplay_io.o \
 			network/netplay/netplay_sync.o \
-			network/netplay/netplay_common.o \
 			network/netplay/netplay_discovery.o \
 			network/netplay/netplay_buf.o
