+++ griffin/griffin.c
@@ -881,12 +881,14 @@ THREAD
 NETPLAY
 ============================================================ */
 #ifdef HAVE_NETWORKING
+#include "../network/netplay/netplay_delta.c"
 #include "../network/netplay/netplay_frontend.c"
+#include "../network/netplay/netplay_handshake.c"
 #include "../network/netplay/netplay_init.c"
 #include "../network/netplay/netplay_io.c"
 #include "../network/netplay/netplay_sync.c"
-#include "../network/netplay/netplay_common.c"
 #include "../network/netplay/netplay_discovery.c"
+#include "../network/netplay/netplay_buf.c"
 #include "../libretro-common/net/net_compat.c"
 #include "../libretro-common/net/net_socket.c"
 #include "../libretro-common/net/net_http.c"
