+++ network/netplay/netplay_delta.c
@@ -1,0 +1,51 @@
+/*  RetroArch - A frontend for libretro.
+ *  Copyright (C) 2010-2014 - Hans-Kristian Arntzen
+ *  Copyright (C) 2011-2016 - Daniel De Matteis
+ *  Copyright (C)      2016 - Gregor Richards
+ *
+ *  RetroArch is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  RetroArch is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with RetroArch.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <stdlib.h>
+#include <sys/types.h>
+
+#include <boolean.h>
+#include <encodings/crc32.h>
+
+#include "netplay_private.h"
+
+bool netplay_delta_frame_ready(netplay_t *netplay, struct delta_frame *delta, uint32_t frame)
+{
+   void *remember_state;
+   if (delta->used)
+   {
+      if (delta->frame == frame) return true;
+      if (netplay->other_frame_count <= delta->frame)
+      {
+         /* We haven't even replayed this frame yet, so we can't overwrite it! */
+         return false;
+      }
+   }
+   remember_state = delta->state;
+   memset(delta, 0, sizeof(struct delta_frame));
+   delta->used = true;
+   delta->frame = frame;
+   delta->state = remember_state;
+   return true;
+}
+
+uint32_t netplay_delta_frame_crc(netplay_t *netplay, struct delta_frame *delta)
+{
+   if (!netplay->state_size)
+      return 0;
+   return encoding_crc32(0L, (const unsigned char*)delta->state, netplay->state_size);
+}
