+++ network/netplay/netplay_init.c
@@ -228,96 +228,6 @@ static bool init_socket(netplay_t *netplay, void *direct_host, const char *serve
    return true;
 }
 
-#ifndef HAVE_SOCKET_LEGACY
-/* Custom inet_ntop. Win32 doesn't seem to support this ... */
-void netplay_log_connection(const struct sockaddr_storage *their_addr,
-      unsigned slot, const char *nick)
-{
-   union
-   {
-      const struct sockaddr_storage *storage;
-      const struct sockaddr_in *v4;
-      const struct sockaddr_in6 *v6;
-   } u;
-   const char *str               = NULL;
-   char buf_v4[INET_ADDRSTRLEN]  = {0};
-   char buf_v6[INET6_ADDRSTRLEN] = {0};
-   char msg[512];
-
-   msg[0] = '\0';
-
-   u.storage = their_addr;
-
-   switch (their_addr->ss_family)
-   {
-      case AF_INET:
-         {
-            struct sockaddr_in in;
-
-            memset(&in, 0, sizeof(in));
-
-            str           = buf_v4;
-            in.sin_family = AF_INET;
-            memcpy(&in.sin_addr, &u.v4->sin_addr, sizeof(struct in_addr));
-
-            getnameinfo((struct sockaddr*)&in, sizeof(struct sockaddr_in),
-                  buf_v4, sizeof(buf_v4),
-                  NULL, 0, NI_NUMERICHOST);
-         }
-         break;
-      case AF_INET6:
-         {
-            struct sockaddr_in6 in;
-            memset(&in, 0, sizeof(in));
-
-            str            = buf_v6;
-            in.sin6_family = AF_INET6;
-            memcpy(&in.sin6_addr, &u.v6->sin6_addr, sizeof(struct in6_addr));
-
-            getnameinfo((struct sockaddr*)&in, sizeof(struct sockaddr_in6),
-                  buf_v6, sizeof(buf_v6), NULL, 0, NI_NUMERICHOST);
-         }
-         break;
-      default:
-         break;
-   }
-
-   if (str)
-   {
-      snprintf(msg, sizeof(msg), msg_hash_to_str(MSG_GOT_CONNECTION_FROM_NAME),
-            nick, str);
-      runloop_msg_queue_push(msg, 1, 180, false);
-      RARCH_LOG("%s\n", msg);
-   }
-   else
-   {
-      snprintf(msg, sizeof(msg), msg_hash_to_str(MSG_GOT_CONNECTION_FROM),
-            nick);
-      runloop_msg_queue_push(msg, 1, 180, false);
-      RARCH_LOG("%s\n", msg);
-   }
-   RARCH_LOG("%s %u\n", msg_hash_to_str(MSG_CONNECTION_SLOT),
-         slot);
-}
-
-#else
-void netplay_log_connection(const struct sockaddr_storage *their_addr,
-      unsigned slot, const char *nick)
-{
-   char msg[512];
-
-   msg[0] = '\0';
-
-   snprintf(msg, sizeof(msg), msg_hash_to_str(MSG_GOT_CONNECTION_FROM),
-         nick);
-   runloop_msg_queue_push(msg, 1, 180, false);
-   RARCH_LOG("%s\n", msg);
-   RARCH_LOG("%s %u\n",
-         msg_hash_to_str(MSG_CONNECTION_SLOT), slot);
-}
-
-#endif
-
 static bool netplay_init_socket_buffers(netplay_t *netplay)
 {
    /* Make our packet buffer big enough for a save state and frames-many frames
